orchestrator/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pymysql==1.1.0
tabulate==0.9.0
flask==3.0.0
flask-cors==4.0.0
openpyxl==3.1.2
